+++ src/transport/request/cart_shopping_dto.go
@@ -1,14 +1,14 @@
 package request
 
 type CartShoppingGetShippingRateDTO struct {
-	StoreID       uint64
-	UserID        uint64
+	StoreID       uint
+	UserID        uint
 	DistributorID string
 }
 
 type CartShoppingCheckoutDTO struct {
-	StoreID                 uint64
-	UserID                  uint64
+	StoreID                 uint
+	UserID                  uint
 	DistributorID           string
 	ShippingProviderService string `json:"shipping_provider_service" validate:"required"`
 	ShippingProviderCode    string `json:"shipping_provider_code" validate:"required"`
